schema: move task deleted migration to 3.1
