+++ pkg/schema/migrate/v3.0.0_integration_test.go
@@ -116,61 +116,3 @@ func TestRewriteHealthCheck30(t *testing.T) {
 
 	testCallback(t, migrate.CallComment, "rewriteHealthCheck30", prepare, validate)
 }
-
-func TestSetExistingTasksDeletedIntegration(t *testing.T) {
-	type Task struct {
-		ClusterID uuid.UUID
-		Type      string
-		ID        uuid.UUID
-		Name      string
-		Deleted   bool
-	}
-
-	clusterID := uuid.MustRandom()
-	input := []*Task{
-		{
-			ClusterID: clusterID,
-			Type:      "repair",
-			ID:        uuid.MustParse("6D77D539-1D10-4A40-98F2-4DE0865DE183"),
-		},
-		{
-			ClusterID: clusterID,
-			Type:      "backup",
-			ID:        uuid.MustParse("8042889C-6226-4FBA-BEAD-13C768FC2FC8"),
-		},
-	}
-
-	prepare := func(session gocqlx.Session) error {
-		q := qb.Insert("scheduler_task").Columns("cluster_id", "type", "id").Query(session)
-		for _, t := range input {
-			if err := q.BindStruct(t).Exec(); err != nil {
-				return err
-			}
-		}
-		q.Release()
-		return nil
-	}
-
-	validate := func(session gocqlx.Session) error {
-		q := qb.Select("scheduler_task").Where(qb.EqLit("deleted", "false")).AllowFiltering().Query(session)
-		defer q.Release()
-
-		var all []*Task
-		if err := q.Iter().Unsafe().Select(&all); err != nil {
-			t.Fatal(err)
-		}
-
-		ctx := context.Background()
-		for i, v := range all {
-			Logger.Debug(ctx, "Tasks", "pos", i, "task", v)
-		}
-
-		if len(all) != 2 {
-			t.Fatalf("Expected 2 not deleted, got %d", len(all))
-		}
-
-		return nil
-	}
-
-	testCallback(t, migrate.CallComment, "setExistingTasksDeleted", prepare, validate)
-}
