+++ schema/v3.0.0.cql
@@ -10,8 +10,5 @@ ALTER TABLE scheduler_task ADD success_count int;
 ALTER TABLE scheduler_task ADD error_count int;
 ALTER TABLE scheduler_task ADD last_success timestamp;
 ALTER TABLE scheduler_task ADD last_error timestamp;
-ALTER TABLE scheduler_task ADD deleted boolean;
-
--- CALL setExistingTasksDeleted;
 
 ALTER TABLE repair_run ADD host text;
