+++ schema/v3.1.0.cql
@@ -1,0 +1,3 @@
+ALTER TABLE scheduler_task ADD deleted boolean;
+
+-- CALL setExistingTasksDeleted;
